camera_control/requirements.txt
Flask==3.0.1
requests==2.31.0
pandas==2.2.2
flask-restx==1.3.0
opencv-python-headless
redis==5.0.4
Pillow==10.3.0
Flask-Cors == 4.0.0
gevent
uvicorn
flask[async]
imutils
gunicorn